pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openinference-instrumentation-openllmetry"
dynamic = ["version"]
description = "OpenInference OpenLLMetry Instrumentation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9, <3.14"
authors = [
  { name = "OpenInference Authors", email = "oss@arize.com" },
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "opentelemetry-api>=1.22.0",
  "opentelemetry-sdk>=1.22.0",
  "opentelemetry-instrumentation>=1.22.0",
  "opentelemetry-semantic-conventions>=1.22.0",
  "openinference-instrumentation>=0.1.27",
  "openinference-semantic-conventions>=0.1.17",
  "typing-extensions>=4.8.0",
  "wrapt>=1.14.0",
  "grpcio>=1.60.0",
]

[project.optional-dependencies]
instruments = [
  "opentelemetry-instrumentation-openai>=0.35.0",
  "opentelemetry-instrumentation-langchain>=0.35.0",
  "opentelemetry-instrumentation-anthropic>=0.35.0",
]

tests = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "python-dotenv>=1.0.0",
]

[project.entry-points.openinference_instrumentor]
openllmetry = "openinference.instrumentation.openllmetry:OpenLLMetryInstrumentor"

[project.urls]
Homepage = "https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-openllmetry"

[tool.hatch.version]
path = "src/openinference/instrumentation/openllmetry/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/openinference"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
  "tests",
]

[tool.mypy]
strict = true
explicit_package_bases = true
exclude = [
  "examples",
  "dist",
  "sdist",
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "wrapt",
  "openai",
  "langchain",
  "langgraph",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "E501"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
force-single-line = false 